player works now again

+++ YourOwnGame/app/src/main/java/yourowngame/com/yourowngame/classes/actors/Barrier.java
@@ -9,7 +9,7 @@ import android.support.annotation.Nullable;
 import android.view.View;
 
 public class Barrier extends GameObject {
-    public Barrier(double posX, double posY, double speedX, double speedY, int[] img, float rotationDegree, @Nullable String name) {
+    public Barrier(double posX, double posY, double speedX, double speedY, int[] img, int rotationDegree, @Nullable String name) {
         super(posX, posY, speedX, speedY, img, rotationDegree, name);
     }
