+++ YourOwnGame/app/src/main/java/yourowngame/com/yourowngame/classes/actors/Reward.java
@@ -27,7 +27,7 @@ import android.view.View;
 
 public class Reward extends GameObject {
 
-    public Reward(double posX, double posY, double speedX, double speedY, int[] img,float rotationDegree, @Nullable String name) {
+    public Reward(double posX, double posY, double speedX, double speedY, int[] img,int rotationDegree, @Nullable String name) {
         super(posX, posY, speedX, speedY, img, rotationDegree,name);
     }
